pyproject.toml
[project]
name = "hello-memorybank"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-cloud-aiplatform>=1.111.0",
    "google-genai>=1.63.0",
    "python-dotenv>=1.2.1",
]
